refactor:extract query logic to src/table/query.rs
